feat(ui): schedule alarm by pressing set alarm button

+++ app/src/main/java/io/github/kirasok/alarmix/presentation/editor/EditorEvent.kt
@@ -3,5 +3,5 @@ package io.github.kirasok.alarmix.presentation.editor
 import io.github.kirasok.alarmix.domain.model.Alarm
 
 sealed class EditorEvent {
-  data class SetAlarm(val hour: Int, val minute: Int) : EditorEvent()
+  data class SetAlarm(val hour: Long, val minute: Long) : EditorEvent()
 }
